Change expiration type from long to int

+++ src/main/java/xyz/holocons/mc/holdthatchunk/ChunkUnloader.java
@@ -12,7 +12,7 @@ public class ChunkUnloader {
     private class DelayedChunkUnload {
 
         private final ClientboundForgetLevelChunkPacket packet;
-        private final long expiration;
+        private final int expiration;
         private boolean canceled;
         private DelayedChunkUnload next;
